Update WCSimRootOptions.cc
when it attempts to fetch git head commit hash, this leads to failures on the grid. There is a message below that states the tar file should be properly git tagged but as we are copying, building, and making modifications from /home/annie/WCSim/WCSim instead of cloning, I just removed this error message for now.