Change to PMT Tilts
Included the measured tilt angles in `PMTPositions_Scan.txt` and directly applied the tilting. This is to remedy a bug in which the PMT_tilt file was 1. not being read in (due to the file not being in the CMake list, and 2. when the PMT properties from the PMTPositions_scan.txt file were being read in, it wasn't matching the PMT_ID to the created pmt_tilt map correctly.